+++ assets/javascripts/kitten/components/modals/next/index.js
@@ -368,16 +368,18 @@ const ModalTitle = ({ children }) => (
   </Title>
 )
 
-const ModalParagraph = ({ children, withoutMargin, className, align }) => (
+const ModalParagraph = ({ children, withoutMargin, className, align, tag, ...props }) => (
   <StyledParagraph
     modifier="tertiary"
-    style={{ textAlign: align }}
     noMargin
-    tag="p"
-    className={classNames('k-ModalNext__paragraph', className, {
+    tag={tag || 'p'}
+    className={classNames('k-ModalNext__paragraph', className,
+      `k-u-align-${align}`,
+      {
       'k-u-margin-bottom-triple': !withoutMargin,
       'k-u-margin-bottom-quadruple@s-up': !withoutMargin,
     })}
+    {...props}
   >
     {children}
   </StyledParagraph>
@@ -505,7 +507,7 @@ const InnerModal = ({
   let customStyle = {
     '--Modal-headerZIndex': headerZIndex,
   }
-  if (contentCols.length !== {}) {
+  if (contentCols !== {}) {
     for (const [key, value] of Object.entries(contentCols)) {
       customStyle[`--Modal-contentCols--${key}`] = value
     }
@@ -525,7 +527,7 @@ const InnerModal = ({
             {
               'k-ModalNext__content--fullSize': fullSize,
               'k-ModalNext__content--customContentCols':
-                contentCols.length !== {},
+                contentCols !== {},
             },
           ),
           afterOpen: 'k-ModalNext--afterOpen',